pyproject.toml
[project]
name = "mock-api"
version = "0.1.0"
description = "Production-ready FastAPI template for internal and external teams"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic-settings>=2.10.0",
    "python-json-logger>=2.0.7",
    "sqlalchemy>=2.0.41",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.9",
    "greenlet>=3.2.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "httpx>=0.28.1",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
    "mypy>=1.17.1",
    "types-python-dateutil>=2.9.0.20250822",
    "bandit>=1.8.6",
    "pre-commit>=4.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app",
    "tests",
    "alembic",
    "infra",
    "README.md",
    "alembic.ini",
    "docker-compose.yml",
    "Dockerfile",
    ".env.example",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N", "ASYNC", "RUF"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = []
exclude = ["^alembic/"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "alembic"]
skips = ["B101"]
